add DataLayer class
