+++ gpt_mini/modeling/tensorflow_char_gpt.py
@@ -2,10 +2,10 @@ import os
 from typing import Generator, Tuple
 
 import tensorflow as tf
-import tensorflow_datasets as tfds
 
 from gpt_mini.modeling.gpt import Gpt
-from gpt_mini.utility import logger, well_known_paths
+from gpt_mini.utility import logger
+from gpt_mini.utility.data_layer import DataLayer
 from gpt_mini.utility.plot_layer import PlotLayer
 
 log = logger.init("tensorflow_char")
@@ -42,29 +42,6 @@ class Tensorflow_Char_Gpt(Gpt):
         self._params["dim_head"] = (
             self._params["embedding_dim"] // self._params["num_heads"]
         )
-
-    def _get_data(self) -> dict:
-        text_dict = dict()
-        if self._data_source == "local":
-            for split in ["train", "validation", "test"]:
-                local_path = os.path.join(
-                    well_known_paths["DATASETS_DIR"],
-                    f"local-{split}.txt",
-                )
-                try:
-                    with open(local_path, "r") as f:
-                        text_dict[split] = f.read()
-                except:
-                    raise RuntimeError(f"missing {local_path}")
-        elif self._data_source == "shakespeare":
-            dataset = tfds.load("tiny_shakespeare")
-            for split in ["train", "validation", "test"]:
-                for element in dataset.get(split).as_numpy_iterator():
-                    # Extract text from dataset using get()
-                    text_dict[split] = element.get("text").decode("utf-8")
-        else:
-            raise RuntimeError("data_source must be one of {'shakespeare', ...}.")
-        return text_dict
 
     def _tokenize(self, text_dict: dict) -> dict:
         # Construct a character-level vocabulary based on the training set
@@ -255,7 +232,8 @@ class Tensorflow_Char_Gpt(Gpt):
         return context
 
     def train(self) -> None:
-        text_dict = self._get_data()
+        datalayer = DataLayer(data_source=self._data_source)
+        text_dict = datalayer._get_data()
         data_dict = self._tokenize(text_dict)
         training_generator = self._generate_batch(data_dict, split="train")
         validation_generator = self._generate_batch(data_dict, split="validation")
@@ -281,7 +259,8 @@ class Tensorflow_Char_Gpt(Gpt):
         self._save_model(model)
 
     def score(self) -> None:
-        text_dict = self._get_data()
+        datalayer = DataLayer(data_source=self._data_source)
+        text_dict = datalayer._get_data()
         data_dict = self._tokenize(text_dict)
         test_generator = self._generate_batch(data_dict, split="test")
